ubuntu 20 basics support for apache, php, hetzner-floatip; fix idempotency at prometheus networks;

+++ hetzner/floatip-bind.sh
@@ -15,27 +15,50 @@ if ! test -z "${2}"; then
 fi
 
 if [ ${OS} = "Ubuntu" ]; then
-  mkdir -p /etc/network/interfaces.d/60-floating-ip.cfg
-  cat >/etc/network/interfaces.d/60-floating-ip.cfg <<EOF
+  if test -f "/etc/lsb-release" && grep -q "Ubuntu 18" "/etc/lsb-release"; then
+    # ubuntu 18
+
+    # todo: check. is/was mkdir necessary?
+    mkdir -p /etc/network/interfaces.d/60-floating-ip.cfg
+    cat >/etc/network/interfaces.d/60-floating-ip.cfg <<EOF
 auto eth0:1
 iface eth0:1 inet static
     address ${FLOATING_IP}
     netmask 32
 EOF
 
-  if ! test -z "${FLOATING_IP6}"; then
-    echo "" >>/etc/network/interfaces.d/60-floating-ip.cfg
-    echo "auto eth0:2" >>/etc/network/interfaces.d/60-floating-ip.cfg
-    echo "    iface eth0:2 inet6 static" >>/etc/network/interfaces.d/60-floating-ip.cfg
-    echo "    address ${FLOATING_IP6}" >>/etc/network/interfaces.d/60-floating-ip.cfg
-    echo "    netmask 64" >>/etc/network/interfaces.d/60-floating-ip.cfg
-  fi
+    if ! test -z "${FLOATING_IP6}"; then
+      echo "" >>/etc/network/interfaces.d/60-floating-ip.cfg
+      echo "auto eth0:2" >>/etc/network/interfaces.d/60-floating-ip.cfg
+      echo "    iface eth0:2 inet6 static" >>/etc/network/interfaces.d/60-floating-ip.cfg
+      echo "    address ${FLOATING_IP6}" >>/etc/network/interfaces.d/60-floating-ip.cfg
+      echo "    netmask 64" >>/etc/network/interfaces.d/60-floating-ip.cfg
+    fi
 
-  read -p " Restart Network? (y|N) " do_restart
+    read -p " Restart Network? (y|N) " do_restart
 
-  if [[ ${do_restart} == "y" ]] || [[ ${do_restart} == "Y" ]]; then
-    service networking restart
-    echo "Network restarted"
+    if [[ ${do_restart} == "y" ]] || [[ ${do_restart} == "Y" ]]; then
+      service networking restart
+      echo "Network restarted"
+    fi
+  else
+    # ubuntu 20
+    touch /etc/netplan/60-floating-ip.yaml
+    cat >/etc/netplan/60-floating-ip.yaml <<EOF
+network:
+   version: 2
+   ethernets:
+     eth0:
+       addresses:
+       - ${FLOATING_IP}/32
+EOF
+
+    read -p " Restart Network? (y|N) " do_restart
+
+    if [[ ${do_restart} == "y" ]] || [[ ${do_restart} == "Y" ]]; then
+      netplan apply
+      echo "Network restarted"
+    fi
   fi
 
 elif [ ${OS} = "CentOS" ]; then
